fix: tei export for ou type attributes [DHIS2-16720]
